+++ lyc_flutter_project/lib/commissions/model/commission_model.dart
@@ -32,9 +32,9 @@ class BasicInfo with _$BasicInfo {
     required int weight,
     required String topSize,
     required String bottomSize,
-    @Default(0) int postalCode,
-    @Default('') String address,
-    @Default('') String detailAddress,
+    required int postalCode,
+    required String address,
+    required String detailAddress,
     required String text,
     required InfoStyle infoStyle,
     required InfoFit infoFit,
@@ -51,6 +51,9 @@ class BasicInfo with _$BasicInfo {
       weight: 0,
       topSize: '',
       bottomSize: '',
+      postalCode: 0,
+      address: '',
+      detailAddress: '',
       text: '',
       infoStyle: InfoStyle.defaultValue(),
       infoFit: InfoFit.defaultValue(),
